+++ api/internationalization.go
@@ -18,11 +18,13 @@ const (
 
 //GetUserChosenLanguage returns the chosen language passed as a custom header
 func GetUserChosenLanguage(req *http.Request) string {
-	if userlng := req.Header.Get(USER_LANGUAGE); userlng == "" {
-		return ""
-	} else {
-		return userlng
+	userlng := ""
+	if userlng = req.Header.Get(USER_LANGUAGE); userlng == "" {
+		if userlng = GetBrowserPreferredLanguage(req); userlng == "" {
+			userlng = "en"
+		}
 	}
+	return userlng
 }
 
 //GetBrowserPreferredLanguage returns the preferred language extracted from the request browser